bank-agent/pyproject.toml
[project]
name = "bank-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agents>=1.4.0",
    "chainlit>=2.6.9",
    "openai-agents>=0.2.8",
    "python-dotenv>=1.1.1",
]
